Adding admins

+++ src/persistance/entity/bot_admin.rs
@@ -4,8 +4,8 @@ use sea_orm::entity::prelude::*;
 #[sea_orm(table_name = "BotAdmin")]
 pub struct Model {
     #[sea_orm(primary_key, auto_increment = false)]
-    pub id: u32,
-    pub user_id: u64,
+    pub id: i64,
+    pub user_id: i64,
 }
 
 #[derive(Copy, Clone, Debug, EnumIter)]
